requirements.txt
pytesseract
arabic-reshaper
python-bidi
hazm
Pillow
wordcloud
